Pull TestImageDownloadOperation into its own file and adjust naming
